fix: avoid empty CSS classes until Linaria bug is fixed (#647)
Linaria 0.7.0 does not play nicely with empty classes/styled elements, sometimes throwing build errors like `[wyw-in-js] Cannot use 'in' operator to search for 'initialCode' in undefined`. There is an [issue raised](https://github.com/Anber/wyw-in-js/issues/144) and a [PR to fix it](https://github.com/Anber/wyw-in-js/pull/147), but it remains unmerged.

This PR adds a CSS comment to all currently empty classes/styled elements. This appears to avoid the build issues, though given the build issues are intermittent anyway, it's hard to know if this will stand up over time.