fix: import map for types (#804)
### Problem
- When ABX updated to `5.0.0-beta.53`, they encountered issues using
`TagGroupItem`, which TS errors reporting that `children` was not a
valid prop.

### Diagnosis
- When we import a component using the `#src/...` import alias, that
import appears unmodified in the `d.ts` file emitted by `tsc`;
- Currently, the import map has `"types": ["./src/*.js",
"./src/*/index.js"]`, which means types will be resolved via the `./src`
folder, which does not exist in the published NPM artefact;
- This prevents consumers from being able to correctly resolve the type
of the `Tag` component which is imported by `TagGroupItem` via the
import path `#src/core/tag`.

### Fix
- Updated the `types` condition for the import map to resolve via the
`./dist` artefacts. This ensures TypeScript can correctly resolve the
types for `#src/...` imports that appear in `d.ts` files.