chore!: update design tokens (#845)
Design tokens in Figma have been updated. This PR brings those updates
to Elements.

Also scouts the removal of `--fix` for the `lint:css` script as it
currently makes changes we don't want. More work needs to be done on the
stylelint config, but having the autofix option in place is annoying
when working with AI.