fix: CDK scripts not finding Node.js types (#852)
The Vitest upgrade done in #851 caused the CDK scripts to no longer find
Node.js types. This is because Vitest moved the Node.js types from
`vitest` to `vitest/node`. This mean our CDK TS code was relying on
Vitest providing the types for Node. This PR fixes this issue by
updating the CDK scripts TS config to explicitly load the types for
Node.