feat: Make `Combobox.Popup` always a modal dialog (#877)
Substantial refactor of the combobox popup. The drawer and popover
variants are now both modal `<dialog>` elements. There's also a new
"auto" variant that responsively changes to a drawer experience on XS
breakpoints. This is now the default popup variant.

Other changes include:
- renamed the popup folder to popup-dialog. It is now decoupled from the
combobox context. A new combobox-popup module has been added to provide
the combobox context integration layer.
- show/hide terminology changed to open/close to match the use of a
modal dialog.
- removal of the input/listbox filtering when the popup opens (this
happens automatically with `<dialog>`).
- the toggle popup button has been renamed to open popup button because
the popup can't be toggled anymore (because it's modal).

Future changes:
- A close button will be introduced to the drawer experience.
- The search input will be updated with a new "naked" variant that will
be used in the drawer experience.
- Popup positioning options will be added to allow consumers to opt in
to the popup being aligned with the top/bottom, left/right edges of the
combobox (currently, it is always aligned with the top and left edges).