chore: refactor `AtAGlanceCardLink` to surround the whole card content rather than the value (#892)
Refactors the card link for At a Glance cards to provide better
accessibility. Originally, the link was designed to surround only the
value, but this does not provide a good label. It's better to link the
whole card content. We can also use a similar approach for the
selectable cards.