chore!: refactor at-a-glance card components (#895)
Big refactor of the At a Glance card components. Instead of having Card,
CardContent, CardLink and CardRadio, we now have Card, AnchorCard and
ButtonCard. This change was made because the original approach did not
place nicely when needing to integrate with Listbox (because CardContent
was always rendering `<h1>` and `<p>` elements, which cannot be placed
inside `<button>` elements.

As such, the three new Card components all share the same base styles,
but they render them using different semantic elements.