feat: `Listbox` now supports `selectAction="auto"` (#903)
Design would like single-selects to work like `selectAction="select"`
and multi-selects to work like `selectAction="toggle"`. To enable this
out-of-the-box, `Listbox` will now default to a new `auto` select action
that will behave like this based on the kind of select being used. The
existing options can still be used to pin a listbox to a particular
behaviour when needed.