feat: add close button to `Combobox.Popup` and update `SearchInput` (#909)
Adds a close button to the `Combobox.Popup` components. It will be
visible whenever the popup is displayed as a drawer. As part of this
enhancement, the search input is not provided to the popup via a
dedicated prop to allow it to be positioned in a separate element than
the listbox of options.

Also, the `SearchInput` is updated to use the clear icon for the clear
button in all cases, rather than just in the borderless variant.
Further, the clear button's size is now based on the search input's
size.