feat: `Combobox` supports `size` prop (#913)
Adds support for a `size` prop to Combobox that ties together the sizing
of the `Combobox` button with the options.

- Adds `size` prop to `Combobox` and provide it via `ComboboxContext`.
- Adds `size` prop to presentational `ComboboxOption` and consume it
from the context in `ComboboxListboxOption`.
- Consume `size` prop from the context in `ComboboxAutocompleteButton`
and `ComboboxSelectButton`

Also fixes an issue introduced by #909 where the popup has insufficient
top padding when there's no header displayed (no search input and no
close button).