feat!: `FormControl` now uses flex layout (#934)
Updates `FormControl` to use a flex layout. Since it was originally
implemented, we have learnt that it is only the `<legend>` element that
does not participate in a `<fieldset>`'s flex layout. By applying bottom
margin on `FormControl.Label` when it is a legend element, we can use a
flex layout and simplify the spacing of the form control's children.